made changes to improve loader
