+++ src/js/containers/LoaderContainer.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 // Components
-import Loader from '../components/core/Loader';
+import Loader from '../components/core/loader/Loader';
 // Actions
 import * as LoaderActions from '../actions/LoaderActions.js';
